pyproject.toml
[tool.poetry]
name = "kiwi"
version = "0.1.0"
description = ""
authors = ["Jerry <zhangji87@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
SQLAlchemy = "^2.0.19"
pyyaml = "^6.0.1"
pymysql = "^1.1.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.0.281"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
select = ["E", "F", "W", "Q", "COM", "I", "N", "B", "RUF"]
line-length = 100
target-version = "py310"

[tool.ruff.per-file-ignores]
"__init__.py" = ["E402", "F401"]

[tool.ruff.flake8-quotes]
inline-quotes = "single"
